requirements.txt
python-telegram-bot==13.4.1
broadlink==0.17.0
cryptography==3.2 # Required by broadlink lib
